fix(render): fix uint/int type mismatch in glyph fragment shader
kind == GLEDITOR_TAG_KIND_PAGE compared a uint against a bare integer
macro, which GLSL (unlike C++) never implicitly converts -- both the
desktop and ES compilers refused the shader outright, breaking every
program that draws a single glyph. Found while verifying Stage 2 of
design/priority-page-building.md: any xudu E2E test that actually rendered
a frame failed at shader compile time, unrelated to that stage's own
Doc::pages changes. GLEDITOR_TAG_KIND_BEAM/_IMAGE already wrap their macro
in uint(...) at each use site; this matches that existing convention
rather than changing the macro's own format string.

+++ assets/shaders/glyph.frag.glsl
@@ -108,7 +108,7 @@ void main() {
   // normalized x/y coordinates so blank-page picks can resolve to the nearest
   // caret instead of unconditionally snapping to the page start.
   uint kind = vTag.x >> GLEDITOR_TAG_KIND_SHIFT;
-  if (0u != vSolid && kind == GLEDITOR_TAG_KIND_PAGE) {
+  if (0u != vSolid && kind == uint(GLEDITOR_TAG_KIND_PAGE)) {
     outTag = uvec4(vTag.x, uint(clamp(vQuadU, 0.0, 1.0) * 65535.0),
                    uint(clamp(vQuadV, 0.0, 1.0) * 65535.0), 0u);
   } else {
